Split CPubKey and friends to new files
